requirements.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
binaryornot==0.4.4
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2
django-allauth==65.7.0
django-browser-reload==1.18.0
django-chartjs==2.3.0
django-tailwind==4.0.1
dotenv==0.9.9
download==0.3.5
fonttools==4.57.0
html5lib==1.1
idna==3.10
Jinja2==3.1.6
jwt==1.3.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql==0.0.3
mysqlclient==2.2.7
oscrypto==1.3.0
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
pypdf==5.4.0
PyPDF2==3.0.1
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.1
razorpay==1.4.2
reportlab==4.4.0
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
uuid==1.30
webencodings==0.5.1
zopfli==0.2.3.post1
